games-util/pokerth-tracker: use default icon path
Currently,games-util/pokerth-tracker does not have an icon shown in the
taskbar. This is because a custom icon install location is defined in
PokerTH_tracker.pro and PokerTH_tracker.desktop. Since Gentoo uses its
own icon install path, the icon is not found during runtime. The
included patch will change the .desktop file to use a generic icon
path which will allow Gentoo to find the icon correctly. This is done
in a similar way to how the icon is defined in games-board/pokerth [1].
This commit was written, tested, and submitted by Lucas Mitrak.

[1] https://github.com/pokerth/pokerth/blob/stable/pokerth.desktop

Package-Manager: Portage-3.0.20, Repoman-3.0.2
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>